Casted variable types to int

+++ 16-binary_tree_is_perfect.c
@@ -66,5 +66,5 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
 		return (0);
 	}
 
-	return (binary_tree_leaves(tree) == (1 << binary_tree_height(tree)) - 1);
+	return ((int)binary_tree_leaves(tree) == (1 << (int)binary_tree_height(tree)) - 1);
 }
